ENH: change numpy sums to float64 This reduces the overall error in the scattering.  Since this is in numpy and on the CPU I don't expect a large speed change.  Absolute tolerances on 1000 atoms are around 1.4e-06 for CPU vs. GPU comparisons, which should be acceptable.  We need to clean up the code now and also rewrite the comparison tests to limit recalculation of data, which should speed the tests up.

+++ pyiid/experiments/elasticscatter/atomics/cpu_atomics.py
@@ -75,7 +75,7 @@ def atomic_fq(task):
     fq = np.zeros((k_max, qmax_bin), np.float32)
     get_fq(fq, omega, norm)
     del q, d, scatter_array, norm, r, omega
-    return fq.sum(axis=0)
+    return fq.sum(axis=0, dtype=np.float64)
 
 
 def atomic_grad_fq(task):
